+++ TroveTools.NET/ViewModel/TroveLocationViewModel.cs
@@ -1,6 +1,7 @@
 ﻿using log4net;
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -9,7 +10,7 @@ using TroveTools.NET.Model;
 
 namespace TroveTools.NET.ViewModel
 {
-    class TroveLocationViewModel : ViewModelBase<TroveLocation>
+    class TroveLocationViewModel : ViewModelBase<TroveLocation>, IDataErrorInfo
     {
         private static readonly ILog log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
         private DelegateCommand _LaunchFolderCommand;
@@ -19,6 +20,15 @@ namespace TroveTools.NET.ViewModel
         public TroveLocationViewModel(string locationName, string locationPath, bool enabled = true)
             : base(new TroveLocation(locationName, locationPath, enabled)) { }
 
+        public string this[string columnName]
+        {
+            get { return DataObject[columnName]; }
+        }
+
+        public string Error
+        {
+            get { return DataObject.Error; }
+        }
 
         public DelegateCommand LaunchFolderCommand
         {